Added the downloading
Using the library pytubefix I am able to download youtube videos with just a link. I am able to put them in a Music folder.

Currently working on being able to play the music without it opening a media player window. Im attempting to use the pydub library to play this sound but am currently having troubles with it.

Added a setup.bat file that installs all the required dependices and will eventually create a shortcut on the desktop.

Added a start.bat file that just runs the python code so that cmd gets automatically put in the right dir