Started testing with read and write lines
In the Test.py I have experimented with saving data in arrays to a txt file. The way it will work is it will save the playlist with the names of all of the music in the same line. Its the playlist name followed by the songs inside the playlist which the program would be able to find. This is how we are going to store playlists