Started working on json files for playlists
Decided to change the storing system from a textfile to a json file for easy readablity and implementation

Inside the main.py when you create a new playlist it is then going to store all the songs in a seperate file and create a part for it in the json file. Currently working on adding the downloaded songs to the json file.

Also added string_checker () to make sure that the user is entering a vaild prompt

started to comment my code on parts that I believe are finished

To do:
Make the main menu display all your current playlists and also make the playlist menu display all the songs inside that playlist