Adds info to json and is able to read from it
Made it so that when you create a new playlist, inside the json file it adds a new block with the name of the playlist. Inside that block there is a single object called song_count.

Made it so that in the main menu it displays all of the playlists that you currently have which is grabbed from the json file. When you choose one of the playlists it takes you to the corrisponding playlist menu

Inside the playlist menu it now displays all the songs that a playlist has underneath it. It does this by grabbing them from the json file. Currently when you choose the song it does not play.

To do:
Make it so when you choose a song from a playlist it plays
Make you able to delete playlists in the main menu
Make it possible to remove songs, skip songs, rename the playlist and toggle shuffling in the playlist menu