Added Many menus
Added a playlist delete menu, playlist rename menu and a song delete menu. I have slowly been working on these 3 menus and decided to push to github over the last 22 days. A long time I know.

The playlist and song delete menus lets you choose from a list of playlists and songs. Whatever is chosen will delete the folder and json infomation. I believe all input case data has been covered so these two menus are complete.

The playlist rename menu works currently as it renames the folder that the playlist is in and renames the key in the json file that is responsible for the playlist. It doesnt have all possible input case data so this menu is not complete.

All thats left to do is actually playing the songs as the file managment is basically finished.

I plan to have two python files in the future. One that contains all the comments and one for the people that dont want to understand it and just want too use the code