Added support for other OS
Added support for the other OS's so other than windows machines can run it. This is done by using pythons OS library which allows me to go to specific paths that change depending on the OS. Also Added the ability too clear the terminal on most OS.

Made it so you cant add another song if there is already a song with the same name

Fixed new_song() crashing when user inputs incorrect infomation

Fixed up the comments so they are not on every line for certain parts and only explain each block.

Made the code a little bit more readable

Known Issues:
When you delete a song the playlist.json file may mess up when adding new files. To fix this I will have to write a new saving system for the songs in the json file