Fixed Messed up json files from Removing and adding songs + More
After many hours of debugging the solution has been found. When the user deleted a song then added a song afterwards there was the possibilty for the new song to override the old song and just dissapear from the playlist.

Made it so when a new song is added in the json file it fills in the blank numbers infront of the keys. This means it no longer overrides the other songs

Made it so exit actually closes the window

Made it so when you delete a song it deletes it by finding the line that the key is on.

Added some OS support I missed