Started Working on playing music and fixed start.bat
when you choose a song it now begins to play music file. Currently the bug is that when you choose another song the old song does stop but the new song repeats itself over and over.

Made it so when one song finishes the next song autoplays and resets to the start if it is at the end of the song list.