Added Playing sounds functionality
Added the ability to play sounds. This works by creating a new thread so that it can run beside the playlist menu and not hault the whole program. Ive added the time module so that it gives the program time to stop the last thread and create a new one.