Added an automatic ffmpeg installation.
Inside the setup.bat file it opens a powershell script to install ffmpeg automatically using choco. As far as I believe this program now works to install ffmpeg when the user runs the setup shortcut.

Added a shortcut that asks for administrator purposes. This is to allow the ffmpeg be installed.