Made heaps of changes to music playing
Fixed music overlapping. I managed to fix this by making a single thread run instead of creating a new thread every single time. Using global variables I am able to tell the thread what to play when the player chooses and depends on if shuffling is enabled

added song shuffling. This works by using a if statement to check if shuffling is enabled. If shuffling is enabled then it will choose a random number out of the possible songs and play that song.

added skipping songs. This works by stopping the currently playing song and adding 1 to choice. This means it will just restart on the next song thus skipping the song