Added Pausing to songs
Added the ability to pause and resume songs. This works by grabbing the current time when the songs starts and when the song is paused it also grabs the current time. This means that it can work out how much time has passed since then and calculate when to start the audiofile from, creating a pause feature.

Fixed song skipping to cause an error. This was because the choice variable was being used for two seperate things causing it to throw an error between different variable types.

To do:
Fix up my code and make it more readable aswell as get rid of unneeded code. Currently the current code is horrible to read so I need to make it look better aswell as add some comments to explain blocks of code.

Add support for buttons outside of the program. This means setting up a button which you can press outside of the program which would pause the song without having to go into the program and pressing p then enter. This should also be a custimizable button per a json file.