Added shuffling that saves
Added a new file called "Settings.json". Inside this json file it allows you to set weither you want the toggle option to be on or off. This will sinc across all playlists aswell as saving when the program is closed. This settings.json file will be used in the future to add other settings that may be useful.

Fixed a bug where after creating a playlist and attempting to go back to the main menu, it takes you back to the playlist creation menu.