Added external keyboard usage
Updated the readme to explain the keyboard library and keybinds.

Changed the Setup.bat file to uninstall the pytubefix and install it again. This is to fix any issues that may occur when youtube change the way videos play. Added the keyboard library to support the use of keyboard inputs without pressing enter

removed UI_Example because UI has been added and for cleanup of the files

added a keyboard_input function which checks for the input keybind from the Settings.json file.