Added the python file as an exe
Made changes to the python script to detect if the script is running as an exe or python file. If it is running as a exe it changes the path otherwise grabs the current directory.

Added an exe which compiles all the dependices which means you can run the exe as long as it has the right paths